backend/requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
crytpography